third_party/blink/web_tests/platform/generic/fast/forms/label/labels-multiple-sibling-labels-expected.txt

Test that the labels list of a form control contains multiple sibling labels.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check button
PASS element.labels.length is 2
Check input
PASS element.labels.length is 2
Check meter
PASS element.labels.length is 2
Check output
PASS element.labels.length is 2
Check progress
PASS element.labels.length is 2
Check select
PASS element.labels.length is 2
Check textarea
PASS element.labels.length is 2
Check buttonType
PASS element.labels.length is 2
Check checkboxType
PASS element.labels.length is 2
Check colorType
PASS element.labels.length is 2
Check dateType
PASS element.labels.length is 2
Check datetimeType
PASS element.labels.length is 2
Check datetime-localType
PASS element.labels.length is 2
Check emailType
PASS element.labels.length is 2
Check fileType
PASS element.labels.length is 2
Check imageType
PASS element.labels.length is 2
Check monthType
PASS element.labels.length is 2
Check numberType
PASS element.labels.length is 2
Check passwordType
PASS element.labels.length is 2
Check radioType
PASS element.labels.length is 2
Check rangeType
PASS element.labels.length is 2
Check resetType
PASS element.labels.length is 2
Check searchType
PASS element.labels.length is 2
Check submitType
PASS element.labels.length is 2
Check telType
PASS element.labels.length is 2
Check textType
PASS element.labels.length is 2
Check timeType
PASS element.labels.length is 2
Check urlType
PASS element.labels.length is 2
Check weekType
PASS element.labels.length is 2
PASS successfullyParsed is true

TEST COMPLETE
